VERSION: 0.1.1 - Toolkit & TGM Integration done. FEW THINGS TO NOTE
-Skin Toolkit is now working perfectly fine with customiser
- TGM activation is now working perfectly. The notification popup hides
itself when the plugin is activated.
- Toolkit plugin version number changed to 0.3
- Oddly, the old Toolkit repo was not getting updated, The same code
was working fine but when downloading it through Git, it was not
working. Git said no changes are being made to the code. So I created a
new repo and everything is now working perfectly fine. So we will just
use the new repo with the name skin-toolkit-plugin
- Version number bumped to 0.2